implement filterable and natural language sort endpoints

+++ src/main/java/com/backend_torch/Data/Persistence/API/dtos/GenderizeResponse.java
@@ -14,6 +14,6 @@ import lombok.NoArgsConstructor;
 public class GenderizeResponse {
     private String name;
    private String gender;
-   private Double probability;
+   private Float probability;
    private  Long count;
 }
